+++ src/stratified_storage/strata/disk_buffer.rs
@@ -97,7 +97,7 @@ mod tests {
     fn test_disk_buffer_normal_write() {
         let filename = "unittest-diskbuffer1.bin";
         let mut disk_buffer = get_disk_buffer(filename, 3, 50, 10);
-        let example = get_example(vec![1, 2, 3]);
+        let example = get_example(vec![1.0, 2.0, 3.0]);
         let data = serialize(&vec![example; 10]).unwrap();
         for _ in 0..5 {
             disk_buffer.write(&data);
@@ -108,7 +108,7 @@ mod tests {
     fn test_disk_buffer_rw_once() {
         let filename = "unittest-diskbuffer2.bin";
         let mut disk_buffer = get_disk_buffer(filename, 3, 50, 10);
-        let example = get_example(vec![4, 5, 6]);
+        let example = get_example(vec![4.0, 5.0, 6.0]);
         let examples = vec![example; 10];
         let data = serialize(&examples).unwrap();
         let index = disk_buffer.write(&data);
@@ -123,7 +123,7 @@ mod tests {
         let filename = "unittest-diskbuffer3.bin";
         let mut disk_buffer = get_disk_buffer(filename, 3, 50, 10);
         for i in 0..5 {
-            let example = get_example(vec![1, 2, i]);
+            let example = get_example(vec![1.0, 2.0, i as f32]);
             let examples = vec![example; 10];
             let data = serialize(&examples).unwrap();
             let index = disk_buffer.write(&data);
@@ -141,7 +141,7 @@ mod tests {
         let mut disk_buffer = get_disk_buffer(filename, 3, 50, 10);
         let mut inputs = vec![];
         for i in 0..5 {
-            let example = get_example(vec![1, 2, i]);
+            let example = get_example(vec![1.0, 2.0, i as f32]);
             let examples = vec![example; 10];
             let data = serialize(&examples).unwrap();
             let index = disk_buffer.write(&data);
@@ -169,7 +169,7 @@ mod tests {
     fn test_disk_buffer_write_disk_full_panic() {
         let filename = "unittest-diskbuffer6.bin";
         let mut disk_buffer = get_disk_buffer(filename, 3, 50, 10);
-        let example = get_example(vec![1, 2, 3]);
+        let example = get_example(vec![1.0, 2.0, 3.0]);
         let data = serialize(&vec![example; 10]).unwrap();
         for _ in 0..6 {
             disk_buffer.write(&data);
@@ -181,12 +181,12 @@ mod tests {
     fn test_disk_buffer_write_feature_size_mismatch_panic() {
         let filename = "unittest-diskbuffer7.bin";
         let mut disk_buffer = get_disk_buffer(filename, 3, 50, 10);
-        let example = get_example(vec![1, 2, 3, 4]);
+        let example = get_example(vec![1.0, 2.0, 3.0, 4.0]);
         let data = serialize(&vec![example; 10]).unwrap();
         disk_buffer.write(&data);
     }
 
-    fn get_example(features: Vec<u8>) -> ExampleWithScore {
+    fn get_example(features: Vec<f32>) -> ExampleWithScore {
         let label: u8 = 0;
         let example = LabeledData::new(features, label);
         (example, (1.0, 0))
