+++ src/stratified_storage/strata/mod.rs
@@ -125,7 +125,7 @@ mod tests {
         let mut strata = Strata::new(1000, 3, 10, filename);
         for i in 0..100 {
             for k in 0..10 {
-                let t = get_example(vec![0, i, k]);
+                let t = get_example(vec![0.0, i as f32, k as f32]);
                 let mut sender = {
                     if let Some(t) = strata.get_in_queue(k as i8) {
                         t
@@ -140,13 +140,13 @@ mod tests {
         for _ in 0..100 {
             for k in 0..10 {
                 let retrieve = strata.get_out_queue(k as i8).unwrap().recv().unwrap();
-                assert_eq!(k, retrieve.0.feature[2]);
+                assert_eq!(k as f32, retrieve.0.feature[2]);
             }
         }
         remove_file(filename).unwrap();
     }
 
-    fn get_example(features: Vec<u8>) -> ExampleWithScore {
+    fn get_example(features: Vec<f32>) -> ExampleWithScore {
         let label: u8 = 0;
         let example = LabeledData::new(features, label);
         (example, (1.0, 0))
